gold counter
- gold now increments over a set period of time
    - only took me 12 hours! kms

+++ Skyfall 2.0/Skyfall 2.0/Assets/_Scripts/Characters/Players/PlayerController.cs
@@ -28,7 +28,7 @@ public class PlayerController : MonoBehaviour
     private bool shootEnabled;
     private bool specialEnabled;
 
-    public int gold;
+    public float gold;
 
     void Awake()
     {
